# docs_api/swagger.yaml
basePath: /
definitions:
  http_public.ErrorResponse:
    properties:
      error:
        example: not found
        type: string
      status:
        example: 404
        type: integer
    type: object
  http_public.InfoResponse:
    properties:
      defaultTtl:
        example: 172800
        type: integer
      implementation:
        example: go
        type: string
      maxChunkSize:
        example: 10485760
        type: integer
      maxTtl:
        example: 604800
        type: integer
      rangeSupported:
        example: true
        type: boolean
      service:
        example: emule-http-cache
        type: string
      uploadRequiresAuth:
        description: |-
          UploadRequiresAuth is false on a server that takes uploads without a
          credential. It is never a reason for a client to drop the key it has — a
          key is still what authorises DELETE, and the operator may close the
          server again tomorrow.
        example: true
        type: boolean
      version:
        example: 1
        type: integer
    type: object
  http_public.UploadResponse:
    properties:
      expires:
        example: 1755500000
        type: integer
      id:
        example: 87d7f7573b0263fc9faf9ed65cb62841
        type: string
      sha256:
        example: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
        type: string
      size:
        example: 9728016
        type: integer
      url:
        example: http://localhost:8080/v1/chunks/87d7f7573b0263fc9faf9ed65cb62841
        type: string
    type: object
info:
  contact: {}
  description: |-
    Encrypted chunk cache for eMuleQt's HTTP Cache feature.
    An uploader stores one AES-256-CBC blob here and hands its URL to several peers, so a part is uploaded once and fetched many times. The server holds ciphertext it has no key for, and is never told the file hash or the part number. Chunk downloads are unauthenticated by design: the 128-bit chunk id is the capability.
  title: emule-http-cache API
  version: "1.0"
paths:
  /:
    get:
      description: An HTML summary of the endpoints this server exposes. Not part
        of the machine contract — a non-Go backend reimplements /v1/* and nothing
        more.
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "503":
          description: server not installed
          schema:
            type: string
      summary: Server status page
      tags:
      - ops
  /install:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: This implementation's own setup page. It writes the config file,
        then shows the generated API key exactly once alongside a clickable ed2k://
        link that configures eMuleQt in one step. Not part of the machine contract.
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
        "503":
          description: the install was refused; nothing was written
          schema:
            type: string
      summary: Setup page
      tags:
      - install
  /v1/chunks:
    post:
      consumes:
      - application/octet-stream
      description: Stores one AES-256-CBC ciphertext blob and returns the absolute
        URL peers should fetch it from. The URL is the only field a client may use
        — never rebuild it from the id, since a backend may serve blobs from another
        host. Content-Length is required; chunked uploads are refused. The server
        is given no file hash, no part index and no key.
      parameters:
      - description: Bearer <apiKey>. Optional only when the server has open_upload
          on.
        in: header
        name: Authorization
        type: string
      - description: Requested lifetime in seconds, clamped to maxTtl
        in: header
        name: X-Chunk-TTL
        type: integer
      - description: Raw ciphertext
        in: body
        name: chunk
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          headers:
            Location:
              description: The same absolute URL as the url field
              type: string
          schema:
            $ref: '#/definitions/http_public.UploadResponse'
        "400":
          description: empty body, or a body that ended short of its Content-Length
          schema:
            $ref: '#/definitions/http_public.ErrorResponse'
        "401":
          description: bad key, or a missing one on a server that requires it
          schema:
            $ref: '#/definitions/http_public.ErrorResponse'
        "411":
          description: no Content-Length
          schema:
            $ref: '#/definitions/http_public.ErrorResponse'
        "413":
          description: over maxChunkSize
          schema:
            $ref: '#/definitions/http_public.ErrorResponse'
        "429":
          description: daily upload quota exhausted
          schema:
            $ref: '#/definitions/http_public.ErrorResponse'
        "503":
          description: server not installed
          schema:
            $ref: '#/definitions/http_public.ErrorResponse'
        "507":
          description: storage failure, or free space below min_free_bytes
          schema:
            $ref: '#/definitions/http_public.ErrorResponse'
      summary: Store a chunk
      tags:
      - chunks
  /v1/chunks/{id}:
    delete:
      description: Only the uploader's key works. A chunk belonging to another key
        reports 404, not 403, so a valid key cannot be used to probe the id space.
        Nobody can authenticate as the reserved "anonymous" owner, so an open server's
        chunks are undeletable by design and lapse at their TTL. A 204 guarantees
        the chunk is gone; a backend that could not remove it says so rather than
        let the client believe a still-downloadable chunk was deleted.
      parameters:
      - description: Chunk id, 32 lowercase hex characters
        in: path
        name: id
        required: true
        type: string
      - description: Bearer <apiKey>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: the chunk is gone
        "401":
          description: bad or missing API key
          schema:
            $ref: '#/definitions/http_public.ErrorResponse'
        "404":
          description: unknown, expired, already deleted, or owned by another key
          schema:
            $ref: '#/definitions/http_public.ErrorResponse'
        "500":
          description: the chunk could not be removed and is still on disk
          schema:
            $ref: '#/definitions/http_public.ErrorResponse'
        "503":
          description: server not installed
          schema:
            $ref: '#/definitions/http_public.ErrorResponse'
      summary: Delete a chunk
      tags:
      - chunks
    get:
      description: 'Unauthenticated by design: the 128-bit id is the capability, and
        the body is ciphertext the server cannot read. Requiring a key here would
        mean sharing the uploader''s credential with every downloader. Single-range
        requests are answered with a 206; a multi-range request gets the whole entity,
        which RFC 9110 section 14.2 permits and which is what the eMuleQt downloader
        can actually parse. HEAD behaves identically without a body.'
      parameters:
      - description: Chunk id, 32 lowercase hex characters
        in: path
        name: id
        required: true
        type: string
      - description: Single byte range, e.g. bytes=0-9728015
        in: header
        name: Range
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: the whole ciphertext
          headers:
            ETag:
              description: First 32 hex characters of the ciphertext digest
              type: string
            X-Chunk-Expires:
              description: Unix time the chunk lapses
              type: string
          schema:
            type: string
        "206":
          description: the requested span
          headers:
            Content-Range:
              description: bytes <first>-<last>/<total>
              type: string
          schema:
            type: string
        "404":
          description: unknown, expired, or a malformed id
          schema:
            $ref: '#/definitions/http_public.ErrorResponse'
        "416":
          description: range not satisfiable
          schema:
            $ref: '#/definitions/http_public.ErrorResponse'
        "503":
          description: server not installed
          schema:
            $ref: '#/definitions/http_public.ErrorResponse'
      summary: Fetch a chunk
      tags:
      - chunks
  /v1/info:
    get:
      description: Unauthenticated. A client must see "service":"emule-http-cache"
        and a version it understands before using the endpoint; this is what stops
        a configuration link pointing it at an arbitrary host. The probe deliberately
        carries no credential.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_public.InfoResponse'
        "503":
          description: server not installed
          schema:
            $ref: '#/definitions/http_public.ErrorResponse'
      summary: Server limits and handshake
      tags:
      - ops
swagger: "2.0"